Revert "Epic/5571 bot preview (#2211)"
This reverts commit 745bea74d07dd002e350ed21baa91418f199349f.
