--- localization/react-intl/src/app/components/cds/chat/Message.json
@@ -1,7 +1,0 @@
-[
-  {
-    "id": "message.unsupportedMessage",
-    "description": "Fallback used in the conversation history when a type of message is not supported",
-    "defaultMessage": "Unsupported message"
-  }
-]
