--- localization/react-intl/src/app/components/cds/mockups/DeviceMockupComponent.json
@@ -1,7 +1,0 @@
-[
-  {
-    "id": "deviceMockup.devicesTooltip",
-    "description": "Tooltip describing that you can change the width of the preview from phone to tablet",
-    "defaultMessage": "Toggle Device Preview Width"
-  }
-]
