+++ src/app/components/drawer/DrawerNavigation.js
@@ -27,12 +27,6 @@ const DrawerNavigation = (props) => {
   const [drawerOpen, setDrawerOpen] = React.useState(getBooleanPref('drawer.isOpen', true));
   const [drawerType, setDrawerType] = React.useState('tipline');
 
-  React.useEffect(() => {
-    if (drawerType === 'bot') {
-      setDrawerOpen(false);
-    }
-  }, [drawerType]);
-
   if (props.teamSlug) {
     const { teamSlug } = props;
